\name{Model-class}
\Rdversion{1.1}
\docType{class}
\alias{Model-class}
\alias{ModelList-class}
\alias{ModelList}

\title{Class \code{"Model"} and \code{"ModelList"}}
\description{
This is a virtual class for generative models used for simulating variables in
SimDat. 
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "Model" in the signature.
}
%\references{
%%  ~~put references to the literature/web site here~~
%}
\author{
Maarten Speekenbrink
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
\examples{
showClass("Model")
}
\keyword{classes}
