\name{UniformModel}
\Rdversion{1.1}
\alias{UniformModel}
\alias{UniformModel-class}
\alias{sd,UniformModel-method}
\alias{mean,UniformModel-method}
\alias{UN}

\title{Uniform model}

\description{
Constructs a Uniform distribution model.
}

\usage{
UniformModel()
}

\details{The UniformModel function constructs a UniformModel.}

\value{An object of class \code{UniformModel} (directly extending the \linkS4class{Model} class).}

\author{Maarten Speekenbrink}
\examples{
mod <- UniformModel()
}

